# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ProConnect professional networking and remote jobs platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profiles:
    get:
      operationId: listProfiles
      tags: [profiles]
      summary: List profiles
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of profiles
          content:
            application/json:
              schema:
                type: object
                properties:
                  profiles:
                    type: array
                    items:
                      $ref: "#/components/schemas/Profile"
                  total:
                    type: integer
                required: [profiles, total]
    post:
      operationId: createProfile
      tags: [profiles]
      summary: Create a new profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfileBody"
      responses:
        "201":
          description: Profile created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profiles/{id}:
    get:
      operationId: getProfile
      tags: [profiles]
      summary: Get profile by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileFull"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: updateProfile
      tags: [profiles]
      summary: Update a profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    delete:
      operationId: deleteProfile
      tags: [profiles]
      summary: Delete a profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /profiles/{profileId}/skills:
    get:
      operationId: listProfileSkills
      tags: [skills]
      summary: List skills for a profile
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of skills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Skill"
    post:
      operationId: addProfileSkill
      tags: [skills]
      summary: Add skill to profile
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSkillBody"
      responses:
        "201":
          description: Skill added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
  /profiles/{profileId}/skills/{id}:
    delete:
      operationId: deleteProfileSkill
      tags: [skills]
      summary: Remove skill from profile
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /profiles/me/privacy:
    get:
      operationId: getMyPrivacy
      tags: [profiles]
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: Owner privacy and source state
          content: { application/json: { schema: { $ref: "#/components/schemas/ProfilePrivacy" } } }
    put:
      operationId: updateMyPrivacy
      tags: [profiles]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ProfilePrivacyUpdate" } } }
      responses:
        "200":
          description: Updated privacy state
          content: { application/json: { schema: { $ref: "#/components/schemas/ProfilePrivacy" } } }
...
components:
  schemas:
    ProfilePrivacy:
      type: object
      required: [privacySettings, discoveryEnabled, sourceMetadata]
      properties:
        privacySettings:
          type: object
          additionalProperties: { type: string, enum: [public, hmr, private] }
        discoveryEnabled: { type: boolean }
        sourceMetadata:
          type: object
          additionalProperties:
            type: object
            properties:
              connected: { type: boolean }
              url: { type: string }
              label: { type: string }
    ProfilePrivacyUpdate:
      type: object
      properties:
        privacySettings: { $ref: "#/components/schemas/PrivacySettings" }
        discoveryEnabled: { type: boolean }
        sourceMetadata: { type: object, additionalProperties: { type: object } }
    PrivacySettings:
      type: object
      additionalProperties: { type: string, enum: [public, hmr, private] }
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Profile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        headline:
          type: string
        bio:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        coverUrl:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        linkedinUrl:
          type: string
          nullable: true
        githubUrl:
          type: string
          nullable: true
        twitterUrl:
          type: string
          nullable: true
        openToWork:
          type: boolean
        accountType:
          type: string
          nullable: true
        industry:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        customLinks:
          type: array
          items:
            type: object
            properties:
              label:
                type: string
              url:
                type: string
            required: [label, url]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, headline, openToWork, createdAt, updatedAt]
    ProfileFull:
      allOf:
        - $ref: "#/components/schemas/Profile"
        - type: object
          properties:
            education:
              type: array
              items:
                $ref: "#/components/schemas/Education"
            experience:
              type: array
              items:
                $ref: "#/components/schemas/Experience"
            portfolio:
              type: array
              items:
                $ref: "#/components/schemas/PortfolioProject"
            skills:
              type: array
              items:
                $ref: "#/components/schemas/Skill"
          required: [education, experience, portfolio, skills]
    CreateProfileBody:
      type: object
      properties:
        name:
          type: string
        headline:
          type: string
        bio:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        coverUrl:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        linkedinUrl:
          type: string
          nullable: true
        githubUrl:
          type: string
          nullable: true
        twitterUrl:
          type: string
          nullable: true
        openToWork:
          type: boolean
      required: [name, headline]
    UpdateProfileBody:
      type: object
      properties:
        name:
          type: string
        headline:
          type: string
        bio:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        coverUrl:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        linkedinUrl:
          type: string
          nullable: true
        githubUrl:
          type: string
          nullable: true
        twitterUrl:
          type: string
          nullable: true
        openToWork:
          type: boolean
        industry:
          type: string
          nullable: true
        glassdoorUrl:
          type: string
          nullable: true
        wellfoundUrl:
          type: string
          nullable: true
        angellistUrl:
          type: string
          nullable: true
        indeedUrl:
          type: string
          nullable: true
        customLinks:
          type: array
          items:
            type: object
            properties:
              label:
                type: string
              url:
                type: string
            required: [label, url]
    Education:
      type: object
      properties:
        id:
          type: integer
        profileId:
          type: integer
        school:
          type: string
        degree:
          type: string
          nullable: true
        fieldOfStudy:
          type: string
          nullable: true
        startYear:
          type: integer
        endYear:
          type: integer
          nullable: true
        description:
          type: string
          nullable: true
      required: [id, profileId, school, startYear]
...
    Experience:
      type: object
      properties:
        id:
          type: integer
        profileId:
          type: integer
        company:
          type: string
        title:
          type: string
        location:
          type: string
          nullable: true
        remote:
          type: boolean
        startDate:
          type: string
        endDate:
          type: string
          nullable: true
        current:
          type: boolean
        description:
          type: string
          nullable: true
      required: [id, profileId, company, title, remote, startDate, current]
...
    PortfolioProject:
      type: object
      properties:
        id:
          type: integer
        profileId:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        projectUrl:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        featured:
          type: boolean
        source:
          type: string
        externalId:
          type: string
          nullable: true
        canonicalUrl:
          type: string
          nullable: true
        objectPath:
          type: string
          nullable: true
        mimeType:
          type: string
          nullable: true
        fileSize:
          type: integer
          nullable: true
        visibility:
          type: string
          enum: [public, private]
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, profileId, title, tags, featured, source, visibility, sortOrder, createdAt, updatedAt]
...
    Skill:
      type: object
      properties:
        id:
          type: integer
        profileId:
          type: integer
        name:
          type: string
        level:
          type: string
          nullable: true
      required: [id, profileId, name]
    CreateSkillBody:
      type: object
      properties:
        name:
          type: string
        level:
          type: string
          nullable: true
      required: [name]
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        HMAC-signed token in the format "profileId:hmac-sha256(profileId, SESSION_SECRET)".
        Returned from /auth/login and /auth/register as authToken.